Added api to context
